scripts/scraper/requirements.txt
playwright==1.49.0
supabase==2.10.0
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
openai>=1.0.0
